remove RecoverTableStmt
