+++ fe/fe-core/src/main/cup/sql_parser.cup
@@ -2483,10 +2483,6 @@ recover_stmt ::=
     {:
         RESULT = new RecoverDbStmt(dbName, dbId, alias);
     :}
-    | KW_RECOVER KW_TABLE table_name:dbTblName opt_id:tableId opt_alias:alias
-    {:
-        RESULT = new RecoverTableStmt(dbTblName, tableId, alias);
-    :}
     | KW_RECOVER KW_PARTITION ident:partitionName opt_id:partitionId opt_alias:alias KW_FROM table_name:dbTblName
     {:
         RESULT = new RecoverPartitionStmt(dbTblName, partitionName, partitionId, alias);
